backend/requirements.txt
asgiref==3.7.2
build==0.10.0
certifi==2023.7.22
Click==7.0
distlib==0.3.7
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
et-xmlfile==1.1.0
filelock==3.12.2
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
packaging==23.1
pipenv==2023.10.24
platformdirs==3.11.0
pyproject_hooks==1.0.0
sqlparse==0.4.4
virtualenv==20.24.6
Werkzeug==0.14.1
